requirements.txt
GitPython~=3.1.45
setuptools~=80.9.0
textual~=6.11.0
black
pre-commit